Rework the logic for creating Makefiles slightly: FreeBSD make(1) (which I use) and NetBSD make(1) (which is what most non-FreeBSD users will need if they want to run 'make publish') have different ideas of 'make -V' means: FreeBSD expands variables, while NetBSD doesn't.  This means we need to either avoid make -V or be very careful in how we use it.
For per-file compile options, I'm avoiding -V, by switching from
having a variable set to the option in question to having a target
which echos the option; this has the side benefit of matching the
way the source-*.o targets operate.

For MAN1, I'm forcing the variable expansion by using := instead
of =.  In this case FreeBSD attempting to expand variables again
has no effect because there is nothing more to expand.