Run coverity only on job X.1
Prior to this commit, the Coverity scanner would attempt to run once
for every build job -- i.e., 4 times, for each combination of OS and
compiler.  This caused problems and wasn't very useful anyway.

With this commit we check if the Travis CI job number is something.1
and then invoke Coverity manually rather than via the addons system.