Fix 'docker stop' issue
Docker runs processes inside a container with an isolated process tree.
Usually your daemon is the only process inside this container. This means
that it receives PID 1, being effectively init. This leads to the default
signal handlers of SIGTERM and SIGINT being ignored, which in turn causes
very slow stops of a container running spiped (it waits for the SIGKILL).

This commit explicitely bind a signal handler to SIGTERM and SIGINT if
spiped detects it is PID 1.