Don't require unused cpu-feature detection code.
Or in more than 50 characters: If we do not detect (or declare) that we
have compiler support for a CPU feature, remove the requirement for the
run-time detection support to be compiled and linked in.

This is done by horrifically abusing the preprocessor and the rules for
declaring objects and their linkage: We declare objects with names
blahblahblah_CPUSUPPORT_FOO_BAR as static, then subsequently declare
as extern, and use, objects with names blahblahblah_X where "X" is
either "CPUSUPPORT_FOO_BAR" or "1" based on whether CPUSUPPORT_FOO_BAR
has been previously defined to 1.

If the compiler supports the CPU feature, this results in symbols being
emitted for blahblahblah_1 which are declared via CPUSUPPORT_FEATURE_DECL
in the corresponding cpusupport_foo_bar.c file.

If the compiler does not support the CPU feature, we get two declarations
for each object named blahblahblah_CPUSUPPORT_FOO_BAR, with the first
marked "static" and the second marked "extern"... but the C standard says
that (6.2.2 paragraph 4) if an identifier is declared "extern" but a
prior declaration specified internal linkage, the second declaration
adopts the (internal) linkage from the first definition -- thereby making
sure that we do not emit symbols which will end up being unresolvable by
the linker.

You are not expected to understand this.