Remove SSE2 cpusupport bits again.
Thanks to the "sparse" support in the new cpusupport bits, all we have
to do this time is remove the cpusupport-[Xx]86-[Ss][Ss][Ee]2.c files;
the cpusupport.h header and cpusupport.sh script now ignore features
for which the compiler-support-detection file does not exist.

This means that the libcperciva in spiped can be a subset of *unmodified*
files in the libcperciva repository, which makes keeping it in sync much
easier.