Silence warnings re unused declarations in macros
In cpusupport.h and elasticarray.h we define macros which in turn
define a large number of variables or functions  -- many of which will
not be used in any particular C file.  While most compilers have some
support for skipping "defined but not referenced" warnings, they are
not smart enough to consistently skip these; ironically, prior to this
commit gcc 4.8 issued warnings about cpusupport but not elasticarray,
while clang 3.6 issue warnings about elasticarray but not cpusupport.

We use the "explicitly reference and cast to void" method of silencing
"unused" warnings, with a trick: Because we're dealing with a macro
which will be used at the top level of a C file, we can't just throw
code in without an enclosing function scope.  So we add a new "dummy"
function -- which itself needs to be referenced in order to avoid a
new "unused function" warning being emitted.  Adding a new function
pointer, which points at the dummy function and is (void) referenced
from the dummy function, ensures that every object which has been
defined is referenced at least once.

Interestingly, while this has no effect on the binaries output by
clang, the dead-object-removal optimizer in gcc 4.8 doesn't seem to
recognize that the references between the dummy function and the dummy
function pointer are circular, and so gcc48 ends up emitting the dead
code into the object file.  As far as I'm concerned, adding ~500 bytes
of dead code with one compiler (which will hopefully be fixed soon)
is a price worth paying to get rid of warnings.