build: use --std=c99 for CPP
I wouldn't believe this if I didn't see it myself, but the cpp that's part of
gcc 4.8.4 fails to find "getopt.h" as a dependency of "main.c" when it uses
the default option of --std=gnu11.  Setting that to --std=c11 or --std=c99 or
--std=c89 produces the correct output.

I chose --std=c99 because POSIX has a c99 utility, and also because this is
only run by developers to create the POSIX-compliant makefiles so I don't see
it as being a portability problem.