spipe: make sure all the I/O has been completed.
spipe exits as soon as the remote socket is closed; however, there might
be data that has been received and passed over to the thread that should
forward it to the local file descriptor, but this thread may not have had
a chance to actually send the data yet.

In pushbits(), store the thread IDs of the forwarding threads.  At the end,
try to pthread_join() all of them, ignoring any errors.