Reuse temporary space in SHA256 routines
A significant amount of temporary space is used by SHA256_Transform
while processing each 64-byte block of input data.  Since the values
stored in this space include the input, we zero those arrays after
we are finished with them.  However, this adds significant overhead,
especially now that we use tricks (see util/insecure_malloc.[ch]) to
try to prevent the compiler from "optimizing away" the zeroing of
temporary values.  (Some other functions also require temporary
space, but not as much and they're not called as often.)

To reduce the sanitization overhead, convert functions into "internal"
versions which are passed a pointer to the temporary space they need;
and create public "wrapper" functions which allocate the temporary
space once, pass it to the internal functions to be used multiple
times, and then zero it once before returning.

This has minimal performance impact on the SHA256 code itself (clang
-O2 loses 2% performance, while gcc -O3 gains 2%; other optimization
levels have < 1% change) but the reduced sanitization of temporary
buffers yields a speedup of 20-30%.