Add conn_list_node to dispatch, pass to callback_conndied()
This does not track connections yet; that is implemented in the next commit.

If all open connections are closed before the spiped server exits, there are
no memory leaks.  If there is at least one open connection, then some memory
is not deallocated; this problem existed before this commit.  Creating (but
not tracking and freeing) a conn_list_node increases the amount of
non-deallocated memory by a few bytes, but that will be fixed in the next two
commits.