Track open connections in dispatch.c
If all open connections are closed before the spiped server exits, there are
no memory leaks.  If there is at least one open connection, then some memory
is not deallocated; this problem existed before this commit.  Creating and
tracking (but not freeing) conn_cookies increases the amount of
non-deallocated memory by a few bytes, but that will be fixed in the next
commit.