POSIX: add check for 'restrict'; defines -std=c99
Compilers are not required to provide the `-std=c99` argument, but it is
supported by gcc, clang, and icc.  If we receive a report about a
compiler/toolchain which does not obey the POSIX standard (i.e. c99) and does
not provide `-std=c99`, then we would need to revisit this fix.