tests-bin: add nc-client utility
This is a vastly simplified version of nc(1).  In particular, it only sends
data from stdin to a local port.  It is only intended for use within the spiped
test suite, as a replacment to nc (since nc is not a POSIX utility and we
therefore cannot rely on its presence or behaviour).