tests/nc-client: clarify client network errors
Changing warn0() to warnp() revealed that when the Travis-CI tests fail, it's
due to a broken pipe.  Definitely a good thing to know!

The libcperciva API states that the callback for network_write() is only called
if there's an error (lenwrit=-1), or we've written at least minwrite bytes.
Since we pass send->nchars as minwrite, we should see that value (unless
there's an error).