Properly quote shell strings
POSIX says that command expansion (e.g., via backticks) takes place
before field splitting; as such,
	export FOO=`echo foo bar`
should have the same effect as
	export FOO=foo bar
i.e., to set FOO=foo and then add the variables FOO and bar (which
will be set to "" if not otherwise initialized) to the environment.

Oddly enough, both bash and FreeBSD's sh get this wrong, and act as
if there are quotes around "FOO=`echo foo bar`".

In the case where the posix-l.sh or posix-cflags.sh shell scripts
emitted multiple options, this resulted in the intended behaviour
being accidentally provided by bash and FreeBSD's sh, while the code
breaks (as it should) in dash (which is POSIX-compliant here).

This commit adds the missing quotes, to ensure that POSIX-compliant
shells give us the intended behaviour.