build: make Makefiles: cpusupport-config.h in subdirs
When running
    make Makefiles
some source files need to find cpusupport-config.h.  For the
"Makefiles" target, we create a fake file in the top-level directory.

Previously, it was hard-coded that the top-level directory was
    ../
which worked fine for spipe/ and spiped/.  By chance, none of the programs
under
    tests/*/
needed cpusupport-config.h, so this wasn't an issue.

However, we now need the flexibility to refer to the top-level directory (in
other projects).  I'm putting it in spiped first, since it's the most mature
project we have which uses Makefile.BSD to produce POSIX-compliant Makefiles.