build: use ${LIBCPERCIVA_DIR} in .../Makefile.BSDs
This does not affect the Makefiles generated by `make Makefiles`.  I'm
proposing it so that it's easier to copy&paste lines from one Makefile.BSD into
another (potentially in a different relative directory).

Unfortunately, we need to define
    LIBCPERCIVA_DIR	=	../libcperciva
in each Makefile.BSD because .PATH.c is a target (not a variable) so we cannot
expand it in Makefile.inc (as we do with IDIRS).