Add autotuning to mpool size
We start with a stack of the specified size, but now keep track of how
many allocation requests we have and how many of them resulted in malloc
being called (because we didn't have any cached allocations).  When we
attempt to free an allocation into the pool and find that the pool is
full, we compare those values; if more than 1/256 of the requests since
the last time the pool was full resulted in malloc calls, we expand the
pool -- hoping that keeping more allocations cached will allow us to
avoid malloc calls in the future.