Avoid SIGSEGV on shutdown with target re-resolution
spiped will segfault on shutdown if target address re-resolution is in
use. This is due to two cases of accessing memory that will be freed by
other parts of the code on the first re-resolution cycle. To avoid this:

The DNS timer has to be marked as expired so it's not freed on shutdown
if already expired. Incidentally, the new instance's is recorded so it
can be cancelled on shutdown if not already expired.

Also, the initial target address list as taken from the command line has
to be marked as under dispatch's purview after dispatch_accept() has
been called.