perftests: add send-zeros
This program essentially does:
    dd if=/dev/zero bs=$BLOCKSIZE count=$COUNT | time nc-client $ADDRESS

Note that we deliberately use monoclock_get() instead of monoclock_get_cpu(),
because we're aiming to benchmark spiped.  The CPU usage of send-zeros is
meaningless; what we care about is the time to send info along the whole chain,
which often includes two spiped servers in separate processes.