crypto_entropy.h: fix comment about buflen
This makes the comment in crypto_entropy.h match the comment in
crypto_entropy.c.  The discrepancy existed in the very earliest version
of this code: spiped 0.9.9.1, as lib/crypto/crypto_entropy.{c,h}.

It's true that the generate() function (inside crypto_entropy.c) can
only take a maximum buflen of 2^16.  This is specified in NIST SP
800-90; section 10.1.2.5 Generating Pseudorandom Bits Using HMAC_DRBG,
which refers to max_number_of_bits_per_request in Table 2 of Section
10.1, which gives a limit of 2^19 bits (which is 2^16 bytes).

However, crypto_entropy_read() is a wrapper around generate().  It takes
care of initializing the PRNG if needed, reseeding when needed, and only
requesting up to 2^16 bytes at once.  As such, there is no limit on the
buflen which is passed to crypto_entropy_read().