tests: add spipe with bps-limited echo server
When cat finishes working its way through the file and spipe reads the
EOF, it will close the _sending_ side of the TCP connection to the
remote spiped process, but the _receiving_ side needs to keep on working
so that we can read the data on its way back.

tests/lorem-send.txt is 446 characters, and we're limiting the echo
server to 1000 bytes per second.  I think that ~500 ms is a good target
for this test, but we can tweak these values if necessary.