spipe/pushbits: use pthread_cleanup_*()
This allows us to use pthread_cancel_thread to cleanly stop a pushbits()
thread.

An alternate approach would be to close() the input socket, but this will fail
if the pushbits() thread is already waiting for the blocking read() command
(which happense almost all of the time).  The only way to interrupt that
read() is pthread_cancel().