tests/pushbits: append NUL to "ping" message
The original code sent 4 characters, "ping".  That wasn't a mistake, and
there was no out-of-bounds memory access.

However, this code:
    malloc(strlen(msg))
does look awfully like somebody forgot that strlen() returns the length
without containg the NUL character.

Since it doesn't matter to the test whether it sends 4 or 5 characters,
I figured that we might as well send the final NUL to avoid the confusion.
So we're now sending "ping\0".

Reported by:	Coverity scan