spipe: cleaner handling of sas_t for error path
Better solution to 6a6fbd977d1c839c0d038ba8d24883c03f080cd7 (2016-08-11);
this is similar to 8faf99d81bd9b9d8e84f47920f21559642b065cd (2017-12-11).

proto_conn_create() is documented as "Free ${sas} once it is no longer
needed", so it's safe to rely on that to free it.

There's no code path in current spipe which leads to a double free, but
this change will lead to cleaner code in the future.