spipe: swap order of initializing threads[0], threads[1]
Once we have both threads initialized, we need to clean up threads[0]
before threads[1], otherwise we might have a race condition which leads
to trying to shutdown() a socket that's already been close()d.

It's easier to keep the clean-up in the error path correct if we
initialize threads[1] first.