tests/valgrind: support two pids from a binary
This allows us to handle forked binaries.

The check_valgrind_basenames() function still considers it an error if
we have more than two logfiles.  In theory, this code should work with
any number of logfiles, but since we're not expecting to have more than
two, this serves as an additional sanity test.