spipe: proto_conn takes over ownership of s[1]
proto_conn_create() stores s[1] and will close it in proto_conn_drop().
This can can happen due to an error in proto_conn, the connection pipe
status changing, or if we manually call it during the error-path cleanup.

As such, main.c should keep track of not owning it, similar to sas_t in
a962e0bc99e9e366940069cdfbc07bf6d12abb30.  Failing to track this
non-ownership did not cause a problem for any code path in current
spipe, but this change will lead to cleaner code in the future.