pushbits: document socket shutdown() in cleanup
We need the shutdown(P->out, SHUT_WR) in workthread_cleanup() to handle
shutting down when the input when we have two pushbits() held together
with a socketpair.  (Checked by `test_pushbits 4`.)

The shutdown() doesn't hurt for the "echo" case (`test_pushbits 2`)
because shutdown() only applies to sockets; this is why we check
for (errno != ENOTSOCK) in workthread_cleanup().