spipe: don't close the input descriptor in pushbits
This stops
    pushbits(STDIN_FILENO, STDOUT_FILENO, &thread)
from automatically closing stdin upon cleanup, as that is probably not
intended.  (Check with `test_pushbits 2`.)

Using pushbits() to echo stdin to stdout is not something that we do in
spipe, so the previous design was not harmful.  But I find it easier to
reason about pushbits() if we add this extra constraint on the API.

This commit adjusts spipe/main.c to accommodate this change.  It does
not change tests/pushbits/main.c; that file will require many more
changes in order to close all open descriptors, and it will be easier to
reason about those changes if they are all done at once.