Increase perftest speed accuracy
I thought that 1 decimal place was enough for the speed, but I'm now
seeing false positives for "outliers" that arise due to this truncation.

I'm confident that 3 decimal places would be enough to resolve this...
but now that we have perftest-plot.py, we can easily see a visual
respresentation and zoom in as much (or little) as we want.  So we might
as well go with 6 decimal places.