valgrind: process each suppression separately
i.e. no more "shove the whole thing through `grep -v` and hope for the best".

In particular, this is needed for an upcoming potential-memleaks
function, which produces the following suppression:

{
   <insert_a_suppression_name_here>
   Memcheck:Leak
   match-leak-kinds: reachable
   fun:calloc
   obj:/lib/libthr.so.3
   obj:/lib/libthr.so.3
   fun:catopen
   fun:strerror_r
   fun:strerror
   fun:pl_workthread_strerror_localtime
   obj:/lib/libthr.so.3
}

If we only omit the "pl_" line (as happens when we rely on `grep -v`),
then the result won't match the one of the leaks from test_warnp on
FreeBSD (arising from strerror() inside a pthread).

This uses the `csplit` utility, which in POSIX:
https://pubs.opengroup.org/onlinepubs/9699919799/utilities/csplit.html