valgrind spiped-servers: wait to allow the logfile to finish writing
In some cases [1], when we kill a backgrounded valgrind spiped server,
the sh scripts will start to grep the logfile before it's finished being
written.  That leads to much confusion, as the framework reports a leak
(or other errors) in a logfile which looks perfectly fine when we
investigate manually (a second or two later, once a human has noticed
the test failure).

[1] maybe 30% - 50% chance per test?  Since it's intermittent, it's hard
to estimate.  (But it's also not worth trying to estimate it, since
there's an easy fix.)

This commit also cleans up a few message streams (printing messages to
stderr rather than stdout), as per 5c3b50e56d5696300a2e58734e2fe5b2b433f82f.