perftests/standalone-enc: use FreeBSD per-source CFLAGS
On FreeBSD, this allows us to do
    make -f Makefile.BSD

This was made possible due to the change in FreeBSD's
/usr/share/mk/bsd.sys.mk in:
    2016-03-25 Implement (ACFLAGS|CFLAGS|CXXFLAGS).SRC globally.
    https://github.com/freebsd/freebsd-src/commit/59f6130d2a153e2628b822d0543479df47d59944

This does not include cpusupport-config.h; so no hardware acceleration
is available when compiling in this manner.

There are subtle differences between various operating systems which
include bmake.  Compiling directly with `make -f MakefileBSD` is not
intended to work on anything other than FreeBSD.  (Regenerating the
POSIX Makefiles should still work on all bmake versions.)