build: share object files
This automagically updates boilerplate in all Makefile.BSDs.

As such, it would temporarily break perftests/standalone-enc (if we ran
`make Makefiles` on the git tree at this point), as it requires a
directory-specific override.  For clarity of the git history, that fix
will happen in the following commit, rather than being smooshed into
this one.