build: remove unnecessary parts due to liball.a
When parts of liball.a are updated, `make` does the right thing and
rebuilds the whole library.

`make -f Makefile.BSD` does depend on liball.a, but it does not perform
the "loop-back" building which would be required to update it.  If you
are compiling with that method and update part of liball.a, you must
manually re-compile the library.