tests/pushbits: add warning about EOF on stdin
In GitHub Actions, when `test_pushbits 2` tries to read from stdin, it
immediately gets an EOF.  As a result, the pushbits workthread has
already stopped by the time the sleep(1) finishes.

(As noted in the comment about ESRCH [1], the thread ID should still be
valid, so the pthread_cancel() should still work, but many OSes don't do
that properly.  The existing warning about ESRCH was already sufficient
for that purpose.)

[1] Added in e56ffd717e9d13c534bbaba3f3515e94e96f4a34.

This new warning is aimed at future developers who look at
echo_stdin_stdout() and expect the thread to still be alive on all
systems, especially when they see the thread alive on all systems they
can manually test on.

In particular, if running the tests inside a docker container, if
docker-run was not called with the --interactive=true flag, then docker
will close stdin.
    https://docs.docker.com/engine/reference/commandline/run/
    (search for "interactive")

GitHub Actions is documented as allowing workflows "directly on the
virtual machine or in a Docker container" [2].  We have not specified a
choice, and I don't see any mention of docker in the test logs, but it's
posible that their non-docker runner also closes stdin.  (I'm guessing
that docker has a good reason for closing stdin by default?  And if so,
that reason would likely apply to github actions as well.)

[2] https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners