mpool.h: add sanity check
M->allocsize is non-zero to begin with, but theoretically if it started
as a power of 2 and we kept on doubling the pool, it could wrap around
to 0 again.  Of course, we would run out of memory long before that
actually occurred.

A more concrete benefit is that some static analysis tools don't realize
that M->allocsize > 0 (despite adding the CTASSERT in the previous
commit), so they warn about a potential malloc(0) in mpool_free().  This
eliminates those spurious warnings.

Reported by:	clang-tidy