perftest.c: avoid shadowing callback function args
The argument names "buf" and "buflen" in callback functions appear to
conflict with the local variables "buf" and "buflen".

This isn't an actual case of shadowing variable names, because the
arguments to callback functions aren't part of perftest_buffers()
itself.  But since those argument names aren't necessary, we might as
well remove them to avoid confusing static analysis programs.

(Note that the function API says what the callback arguments are, so
we're not losing any information.)

Reported by:	cppcheck-2.7.4