scripts: use underscores for "private" functions
In POSIX sh, there's no such thing as "public" or "private" functions.

However, functions are allowed to contain underscores [1], so we're
adopting the convention of leading underscores indicating a function
which is not intended to be called by code in a different file.

[1] https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_235

Private functions in shared_test_functions.sh:
    prepare_directory
    notify_success_or_fail
    scenario_runner

Private functions in shared_valgrind_functions.sh:
    valgrind_prepare_directory
    valgrind_check_optional
    valgrind_process_suppression_file
    valgrind_ensure_suppression
    valgrind_get_basename
    valgrind_check_logfile