Invoke usage for missing hostnames in -[st]
A source or destination address of the form ":1234" is parsed as a
hostname of "" which is guaranteed to never work.  While we can't
reject all possible invalid addresses at this point -- and probably
shouldn't since the resolver can give more informative errors --
this particular case can easily be produced from a broken shell
script (with a hostname variable accidentally not set), so it's more
worthy of rejection via usage() than most.

Suggested by:	Ross Richardson