sock_util: add sock_addr_validate()
A source or destination address of the form ":1234" is parsed as a
hostname of "" which is guaranteed to never work.  While we don't want
to give special error messages for all possible invalid addresses, this
particular case can easily be produced from a broken shell script (with
a hostname variable accidentally not set), so it's worth having a
special error message.

Suggested by:	Ross Richardson