perftests/recv-zeros: more specific error-handling
clang19-tidy warned that the previous code could result in an undefined
value in errno.  That's because in general, recv() can return -1, 0, or
any positive number up to buflen.  However, because we used the
MSG_WAITALL flag, the possible output is restricted to -1, 0, or buflen.
Due to that flag, the previous code was not wrong.

However, there's no harm in changing the "do we have an error?" check,
and that makes clang-tidy happier.