tests/pushbits: clarify message checking
clang-tidy-19 complained that the comparison length was too long and was
a possible buffer overflow.  I think this essentially boiled down to
"what does strncmp() do with non-NUL-terminated arrays?".

My interpretation of C99 is that it's ok -- despite the name, strncmp()
is specified to compare arrays, not strings -- but using memcmp() is
more clear anyway, so we might as well switch.